Split printers into separate files
