+++ lib/spandx/core/printer.rb
@@ -23,48 +23,5 @@ module Spandx
         end
       end
     end
-
-    class CsvPrinter < Printer
-      def match?(format)
-        format.to_sym == :csv
-      end
-
-      def print_line(dependency, io)
-        io.puts(CSV.generate_line(dependency.to_a))
-      end
-    end
-
-    class JsonPrinter < Printer
-      def match?(format)
-        format.to_sym == :json
-      end
-
-      def print_line(dependency, io)
-        io.puts(Oj.dump(dependency.to_h))
-      end
-    end
-
-    class TablePrinter < Printer
-      def match?(format)
-        format.to_sym == :table
-      end
-
-      def print_header(_io)
-        @dependencies = SortedSet.new
-      end
-
-      def print_line(dependency, _io)
-        @dependencies << dependency
-      end
-
-      def print_footer(io)
-        table = Terminal::Table.new(headings: ['Name', 'Version', 'Licenses', 'Location'], output: io) do |t|
-          @dependencies.each do |d|
-            t.add_row d.to_a
-          end
-        end
-        io.puts(table)
-      end
-    end
   end
 end
